Fixing some stuff
Post-merge first fixing. Some problems resolve, but there are still a lot of laks. The most important: it's not important if I select 'Yes' or 'No' in the window which appears when someone wants to send me a file, because in any case a new job will be added in the receiving queue and the zip file will be created; but the operation will stuck on the 0%, and no transmission will be performed.