Interruptions management
The app is now able to manage and show to the user errors occourring during outgoing and incoming transfers.
+ App.xaml.cs and BackgroundForm.cs now cooperate in order to show to the user a balloon tip informing him about errors.
+ The NotificationWindow.xaml.cs (with the support of an intelligent data management of the insertion and remotion events in the FileShareDataContext.cs) is now able to substitute the progress bar with a message in two cases: 1) when an error occours (in this case shows a red 'Error' message); 2) when the trasnfer is completed (in this case shows a green 'Completed' message).
+ A new class ListdJob.cs was added to wrap a job and to contain several additional information about a transfer. The ListView in the NotificationWindow is now an ObservableCollection<ListedJob>: the FileShareDataContext manages the translation from Job to ListedJob and some other features which makes the ListView more user friendly.

In order to obtain these results, several core classes in the FileTransfer and Storage namespaces have been modified:
1) ReceptionExecutor.cs class added. A ReceptionExecutor behaves like a JobExecutor: is "scheduled" by the ServerClass after the Accept and executes the operations of creating a TnsServer and managing the interaction with it, registering some callback to the events the TnSServer exposes.
On the OnRequestReceived event of TnsServer, ReceptionExecutor trigger its own RequestReceived event; the ServerClass have registered a callback on this event and trigger its own event  too, so that the gui part of the application is informed about the request reception and chan show to the user a message to ask him if he wants to accept the file.
On the JobInitialized event of TnSServer (event triggered by TnsServer when the JobZipStorageModule.createJob(...) have been completed), the ReceptionExecutor registers the job and add it in the JobsList.Receiving.
 2) TnSServer doesn't catch exceptions anymore: it throws them up and let the ReceptionExecutor catch them. This class triggers the events described in (1).
3) The JobZipStorageModule.cs class has no more the responsibility of managing the insertion/remotion of a Job in/from the JobsList.Receiving list.

EXCEPTION MANAGING
JobExecutor and ReceptionExecutor are responsible of catching SocketExceptions. They can then trigger an event dedicated to this error situation, so that GUI process can be informed about this situation and behave consequently.
The most important thing to do in JobExecutor and ReceptionExecutor is this one: remove the job from the JobsList.Receiving or JobsList.Sending list.