Transfer interruption (explicit or not) completed
+ JobExecutor and ReceptionExecutor now close sockets even if an ecception occours.
+ TnsClient catch exceptions and always executes the closing of the FileIterator he is responsible of.
A SendTimeout is defined on the socket, so that the Send is never definitively blocking.
+ TnsServer catch exceptions and always executes the closing of the FileIterator he is responsible of.
A ReceiveTimeout is defined on the socket, so that the Send is never definitively blocking.

Extra: Fixing zip storage lack
ZipStorageModule was putting the zip of the sending files in the wrong directory (not the one where file to send was). Behaviour fixed.
Transfer interruption (explicit or not) completed

+ JobExecutor and ReceptionExecutor now close sockets even if an ecception occours.
+ TnsClient catch exceptions and always executes the closing of the FileIterator he is responsible of.
A SendTimeout is defined on the socket, so that the Send is never definitively blocking.
+ TnsServer catch exceptions and always executes the closing of the FileIterator he is responsible of.
A ReceiveTimeout is defined on the socket, so that the Send is never definitively blocking.

Extra: Fixing zip storage lack
ZipStorageModule was putting the zip of the sending files in the wrong directory (not the one where file to send was). Behaviour fixed.