Installer done - To fix to future improvements
+ Installer craetes everything is needed to run correctly the application. The application in installed in a dedicated directory in Programs and will be availale immediately to the user (he only have to run the installer). This means that the application creates the FileChooser application and puts the correct registry key in the registry in order to run it.

NOTE.
The solution is provided in the following form: the FileChooser directory was removed and substituted by a project inside the FileShare main directory (it should never be modified in the future); in the same directory there is the FileShare project. The third project in the FileShare main directory is the FIleShareSetup, which his a project with all its Metadata.
To see everything inside visual studio is needed to create a solution which contains all the three projects. The FileShareSetup will reference the other two projects, but they must stay in the same solution in order to allow this.

IMPORTANT: The FileChooser must be compiled with an Assembly Name equals to FileChooser.exe. Other names will be wrong, and the registry key created by the installer will point to a not existing executable. 
Please pay attention to this when defining the properties of the FileChooser project.