Peer Id, Error mangament, Bug fixing
+ Error management improved. Now different kind of errors are manged and notified to the user. For exemple all the possible error on the ZipArchive are correctly managed and notified, including errors on the length of the file.
To achieve this result, JobExecutor, JobScheduler, ReceiverExecutor and ServerClass were modified in order to be coherent with the Exception hierarchy provided by C# and to propagate the catched exception as argument of the delegate directly to the App.xaml.cs, that can manage in different way different kind of exceptions.
TnSClient now can catch and throw exceptions in its constructor. However before doing this it always close the just opened iterator (if any).
+ Bug on the icon on the ListedPeer fixed. This bug fixing involved an editing in ImageAdapter (check null value).
+ Peer id is now composed only by the name of the peer and it's ip address.
+ FileShareDataContext was behaving in strange way, calling the method to update ListedJob state more than once. Fixed.